Fix backspace in tmux and vim fugitive diff colors
I was wrestling through an issue where backspace was sending the wrong
keycodes to tmux shells. Additionally, vim fugutive diffs did not have
the green color and were falling back to a grey.

Not sure precisely why but after trying several SO posts etc. this
combination of changes fixed things up for me.